fix(ebt): align random noise dtype with embeddings

+++ nova/ebt/modeling_ebt.py
@@ -277,7 +277,11 @@ class EBT_NLP(LightningModule):
         if self.hparams.denoising_initial_condition == "most_recent_embedding":
             raise NotImplementedError(f"most_recent_embedding denoising_initial_condition not supported for NLP yet")
         elif self.hparams.denoising_initial_condition == "random_noise":
-            predicted_tokens = torch.randn(size=(embeddings.shape[0], embeddings.shape[1], self.vocab_size), dtype=torch.bfloat16, device = self.device) * self.hparams.gaussian_random_noise_scaling
+            predicted_tokens = torch.randn(
+                size=(embeddings.shape[0], embeddings.shape[1], self.vocab_size),
+                dtype=self.embeddings.weight.dtype,
+                device=self.device,
+            ) * self.hparams.gaussian_random_noise_scaling
         elif self.hparams.denoising_initial_condition == "zeros":
             predicted_tokens = torch.zeros(size=(embeddings.shape[0], embeddings.shape[1], self.vocab_size), device = self.device)
         else:
